requirements.txt
# requirements.txt

flask
flask-cors
assemblyai
openai
python-dotenv
pydub